+++ _scripts/eleventyBuild.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-echo "Building $PLATFORM to build/$PLATFORM"
-mkdir -p ./build/$PLATFORM
-npx eleventy --output=build/$PLATFORM