+++ _scripts/eleventyWatch.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-echo "Building $PLATFORM to build/$PLATFORM and watching"
-npx eleventy --output=build/$PLATFORM --watch